+++ src/main/java/nextstep/reservation/dto/request/ReservationRequestDto.java
@@ -1,0 +1,27 @@
+package nextstep.reservation.dto.request;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.*;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ReservationRequestDto {
+
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private LocalDate date;
+
+    @JsonFormat(pattern = "HH:mm")
+    private LocalTime time;
+
+    private String name;
+
+    private Long themeId;
+
+}
