+++ src/main/java/nextstep/reservation/dto/response/ThemeResponseDto.java
@@ -1,4 +1,4 @@
-package nextstep.reservation.dto;
+package nextstep.reservation.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -14,7 +14,7 @@ public class ThemeResponseDto {
     private String desc;
     private Integer price;
 
-    public ThemeResponseDto(Theme theme) {
+    public ThemeResponseDto (Theme theme) {
         this.id = theme.getId();
         this.name = theme.getName();
         this.desc = theme.getDesc();
