+++ src/main/java/nextstep/reservation/entity/Theme.java
@@ -4,12 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class Theme {
+    private Long id;
     private String name;
     private String desc;
     private Integer price;
